update theme structure, replace AuthRepository with specific repositories, and add biometric permission
