+++ app/src/main/res/values/themes.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-
-    <style name="Theme.LadyCure" parent="android:Theme.Material.Light.NoActionBar" />
-
+    <style name="Theme.LadyCure" parent="Theme.MaterialComponents.Light.NoActionBar">
+        <item name="colorPrimary">#C281A9</item>
+    </style>
 </resources>
